Move communication and discovery methods to dedicated files
